erp_ai_assistant/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1
sqlalchemy==2.0.41
aiosqlite==0.21.0
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
